requirements.txt
allure-pytest==2.8.13
allure-python-commons==2.8.13
appdirs==1.4.3
argcomplete==1.11.1
attrs==19.3.0
beautifulsoup4==4.9.0
blinker==1.4
Brotli==1.0.7
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9.2
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpath==0.82
kaitaistruct==0.8
ldap3==2.7
lxml==4.5.0
MarkupSafe==1.1.1
mitmproxy==5.1.1
more-itertools==8.2.0
packaging==20.3
passlib==1.7.2
pluggy==0.13.1
prometheus-client==0.8.0
protobuf==3.11.3
publicsuffix2==2.20191221
py==1.8.1
pyasn1==0.4.8
pycparser==2.20
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
pytest==5.4.1
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sortedcontainers==2.1.0
soupsieve==2.0
tornado==6.0.4
urllib3==1.25.8
urwid==2.1.0
userpath==1.3.0
virtualenv==20.0.20
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
wsproto==0.15.0
zstandard==0.13.0
